Change entry point of program to CliParserService
This is to prevent bloating of Duke class
